rename: order, shoppingCart, section 디렉터리 내 관련 컴포넌트 이동
