--- src/components/Header/index.tsx
@@ -1,8 +1,0 @@
-import React from "react";
-import { Wrapper } from "./style";
-
-const Header = () => {
-  return <Wrapper>SHOP</Wrapper>;
-};
-
-export default Header;
