+++ src/components/index.ts
@@ -1,15 +1,17 @@
-import CartDescription from "./CartDescription";
-import CartItemList from "./CartItemList";
-import Header from "./Header";
-import Footer from "./Footer";
-import CartPrice from "./CartPrice";
-import OrderConfirm from "./OrderConfirm";
+import Header from "./section/Header";
+import Footer from "./section/Footer";
+import CartDescription from "./shoppingCart/CartDescription";
+import CartItem from "./shoppingCart/CartItem";
+import CartItemList from "./shoppingCart/CartItemList";
+import CartPrice from "./shoppingCart/CartPrice";
+import OrderConfirm from "./order/OrderConfirm";
 
 export {
-  CartDescription,
-  CartItemList,
   Header,
   Footer,
+  CartDescription,
+  CartItem,
+  CartItemList,
   CartPrice,
   OrderConfirm,
 };
