+++ src/components/section/Header/index.tsx
@@ -1,0 +1,11 @@
+import { Wrapper } from "./style";
+
+interface HeaderProps {
+  children: React.ReactNode;
+}
+
+const Header = ({ children }: HeaderProps) => {
+  return <Wrapper>{children}</Wrapper>;
+};
+
+export default Header;
