+++ src/components/section/Header/style.ts
@@ -1,7 +1,10 @@
 import styled from "styled-components";
 
 export const Wrapper = styled.header`
-  padding: 24px;
+  display: flex;
+  align-items: center;
+  height: 64px;
+  padding-left: 24px;
   background-color: var(--black-color);
   color: var(--white-color);
   font-weight: var(--font-weight-extra-bold);
