+++ src/components/shoppingCart/CartItem/style.tsx
