+++ src/components/shoppingCart/CartPrice/index.tsx
@@ -1,5 +1,5 @@
 import { useRecoilValue } from "recoil";
-import { cartPriceState } from "../../stores/cartPrice";
+import { cartPriceState } from "../../../stores/cartPrice";
 import { Wrapper, PriceKind, PriceNumber, Price } from "./style";
 
 export interface PriceStyleProps {
