+++ src/components/shoppingCart/CartPrice/style.ts
